+++ lbrynet/file_manager/EncryptedFileCreator.py
