+++ lbrynet/file_manager/EncryptedFileStatusReport.py
