+++ lbrynet/lbrynet_daemon/Publisher.py
@@ -5,7 +5,7 @@ import os
 from twisted.internet import defer
 
 from lbrynet.core import file_utils
-from lbrynet.lbryfilemanager.EncryptedFileCreator import create_lbry_file
+from lbrynet.file_manager.EncryptedFileCreator import create_lbry_file
 from lbrynet.lbry_file.StreamDescriptor import publish_sd_blob
