+++ tests/unit/lbrynet_daemon/test_Downloader.py
@@ -12,7 +12,7 @@ from lbrynet.core.StreamDescriptor import StreamDescriptorIdentifier,StreamMetad
 from lbrynet.lbry_file.client.EncryptedFileOptions import add_lbry_file_to_sd_identifier
 from lbrynet.core.HashBlob import TempBlob
 from lbrynet.core.BlobManager import TempBlobManager
-from lbrynet.lbryfilemanager.EncryptedFileDownloader import ManagedEncryptedFileDownloaderFactory
+from lbrynet.file_manager.EncryptedFileDownloader import ManagedEncryptedFileDownloaderFactory
 from lbrynet.lbrynet_daemon.ExchangeRateManager import ExchangeRateManager
 
 from tests.mocks import BlobAvailabilityTracker as DummyBlobAvailabilityTracker
